Rev 3 PCB with gnd hand routed rest autorouted
This PCB is close to the original layout, based on overlay of frontpanel photo (in layout). It does not have space for the onboard keypad and LCD. 10 pin connectors for LCD display and keypad are on the right hand side, next to the Arduino USB connector. Fixes compared to the original V2 are:
- use 1 mm holes for J1 and Arduino
- swapped switch and LED for D13 and D14
- control and data switches are larger (miniature)
- lever UP is 1 for data switches
- longer resistor footprints used for easier mounting
- use normal not wide TTL footprints
- exec and defer corrected on J1

Todo: provide enough mounting holes for frontpanel